Split Error definition into error mod
