+++ src/lapack_binding.rs
@@ -1,32 +1,8 @@
 
 extern crate lapack;
+
 use self::lapack::fortran::*;
-
-use std::error;
-use std::fmt;
-
-#[derive(Debug)]
-pub struct LapackError {
-    return_code: i32,
-}
-
-impl From<i32> for LapackError {
-    fn from(code: i32) -> LapackError {
-        LapackError { return_code: code }
-    }
-}
-
-impl fmt::Display for LapackError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "LAPACK: return_code = {}", self.return_code)
-    }
-}
-
-impl error::Error for LapackError {
-    fn description(&self) -> &str {
-        "LAPACK subroutine returns non-zero code"
-    }
-}
+use error::LapackError;
 
 /// wrapper for *syev functions in LAPACK
 pub trait Eigh: Sized {
